// testace-app/shared/types.d.ts
export interface Question {
    _id: string;
    content: string;
    type: QuestionType;
    subject: string;
    topic: string;
    difficulty: DifficultyLevel;
    options?: string[];
    correctAnswer: string | number;
    explanation: string;
    hints: string[];
    timeLimit?: number;
    tags: string[];
    createdBy: string;
    createdAt: Date;
}
export declare enum QuestionType {
    MULTIPLE_CHOICE = "multiple_choice",
    TRUE_FALSE = "true_false",
    FILL_BLANK = "fill_blank",
    ESSAY = "essay",
    MATH = "math"
}
export declare enum DifficultyLevel {
    EASY = "easy",
    MEDIUM = "medium",
    HARD = "hard"
}